requirements.txt
azure-functions
boto3
botocore